Fix several test failures on x86-64 with -m32.
 - more fixes for crypt-sunmd5.c when 'unsigned long' and 'unsigned
   int' are the same
 - test-getrandom-fallbacks needs to interpose open64 as well as open
 - test-short-outbuf.c can use %zu instead of %lu to avoid casting
 - glibc for x86-64/-m32 uses the same symbol versions as glibc/i386

also, libcrypt.minver is now in alphabetical order by host_cpu pattern
within each block of architectures with the same minimum symbol version.

+++ test-short-outbuf.c
@@ -44,8 +44,8 @@ main (void)
           ok = false;
         }
 
-      printf ("Test %u.0: %s, expected: \"%-2s\", got: \"%-2s\"\n",
-              (unsigned int)(i + 1), result, testcases[i].exp_rn, *outbuf);
+      printf ("Test %zu.0: %s, expected: \"%-2s\", got: \"%-2s\"\n",
+              i + 1, result, testcases[i].exp_rn, *outbuf);
 
       crypt_ra ("@@", "@@", (void **) outbuf, (int *) j);
 
@@ -59,8 +59,8 @@ main (void)
           ok = false;
         }
 
-      printf ("Test %u.1: %s, expected: \"%-2s\", got: \"%-2s\"\n",
-              (unsigned int)(i + 1), result, testcases[i].exp_ra, *outbuf);
+      printf ("Test %zu.1: %s, expected: \"%-2s\", got: \"%-2s\"\n",
+              i + 1, result, testcases[i].exp_ra, *outbuf);
 
       free (j);
       free (*outbuf);
